requirements.txt
pandas
seaborn
numpy
matplotlib
sklearn
scipy
copulas
umap-learn
ctgan
sdv
ydata-synthetic
